lmb-1513: move asyncSome and asyncEvery to a private file
